Fifth-pass: strftime missing, int casts, silent errors, SQLite syntax, frontend types
Datetime decode fixes (5 files):
- apikeys.rs: wrap created_at, last_used_at with strftime
- backups.rs: wrap created_at in list_vm_backups
- snapshots.rs: wrap all 4 UNION branches in list_vm_timeline + list_vm_snapshots
- fleet_automation.rs: wrap last_run_at in both list and re-fetch queries
- migration_jobs.rs: wrap created_at in list_migration_jobs and list_vm_migration_jobs

Integer cast fixes (6 files):
- policy_operator.rs: bind sla_hours as i32 directly (not .to_string()), bind i64 count
- operations.rs: bind vm_count as i64 in both showback inserts
- gitops.rs: bind upserted as i64 (not i32)
- worker.rs: bind vCPU count as i64 (not i32)

Silent error logging (5 files):
- ha.rs: log fence_host failure (split-brain risk) and ha.recover enqueue failure
- drs.rs: log persist_recommendations and vm.migrate enqueue failures
- operations_scheduler.rs: log execute_runbook failure on auto-trigger
- host_os.rs: log guest_ip UPDATE failure

Invalid SQLite syntax / compat (3 files):
- observability.rs: replace OFFSET MAX(...subquery...) with pre-computed Rust value
- zeus_security.rs: replace ->>'title' with json_extract(payload, '$.title') for SQLite <3.38
- infra_graph.rs: replace detail->>'name' with json_extract(detail, '$.name')

Frontend type correctness (6 files):
- ai.ts: fix getFleetSummary to return FleetZeusSummary, getFleetLocal to return FleetClusterSlice
- ai.ts: add task_ids field to runAutopilotSafe results
- platform.ts: fix BackupRecord (remove phantom restore_status, add real backup_path)
- platform.ts: add disk_read_iops/disk_write_iops to getPlatformVmMetrics
- platform.ts: widen retryTask return type to include status and operation
- PlatformZeusOs/MachineFinderBriefing/MissionControlBriefing: construct summary text
  from real FleetZeusSummary fields instead of phantom .summary property

+++ controller/src/engine/operations.rs
@@ -259,7 +259,7 @@ async fn ensure_showback_snapshots(pool: &SqlitePool) -> anyhow::Result<()> {
         )
         .bind(Uuid::new_v4())
         .bind((vm_count as f64) * 12.0)
-        .bind(vm_count as i32)
+        .bind(vm_count)
         .execute(pool)
         .await?;
         return Ok(());
@@ -290,7 +290,7 @@ async fn ensure_showback_snapshots(pool: &SqlitePool) -> anyhow::Result<()> {
         .bind(&name)
         .bind(cost)
         .bind(if vm_count > 5 { "C" } else { grade.as_str() })
-        .bind(vm_count as i32)
+        .bind(vm_count)
         .execute(&mut *tx)
         .await?;
     }
